music-bot/requirements.txt
discord.py>=2.3.0
yt-dlp>=2024.1.1
PyNaCl>=1.5.0
python-dotenv>=1.0.0
